webapp/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0
python-multipart>=0.0.6
itsdangerous>=2.1.0
